tt: restruct project so parsing of cfg is now in configure package
Moved parsing into package configure, to resolve
dependences moved config structs to separate package.

Part of #31
