+++ cli/cmd/check.go
@@ -4,6 +4,7 @@ import (
 	"github.com/apex/log"
 	"github.com/spf13/cobra"
 	"github.com/tarantool/tt/cli/cmdcontext"
+	"github.com/tarantool/tt/cli/configure"
 	"github.com/tarantool/tt/cli/modules"
 	"github.com/tarantool/tt/cli/running"
 )
@@ -26,7 +27,7 @@ func NewCheckCmd() *cobra.Command {
 
 // internalCheckModule is a default check module.
 func internalCheckModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
-	cliOpts, err := modules.GetCliOpts(cmdCtx.Cli.ConfigPath)
+	cliOpts, err := configure.GetCliOpts(cmdCtx.Cli.ConfigPath)
 	if err != nil {
 		return err
 	}
